core/primitives/src/test_utils.rs ressurected
