+++ core/primitives/src/transaction.rs
@@ -5,9 +5,9 @@ use std::hash::{Hash, Hasher};
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize};
 
-use near_crypto::{EmptySigner, PublicKey, Signature, Signer};
+use near_crypto::{PublicKey, Signature};
 
-use crate::account::{AccessKey, AccessKeyPermission};
+use crate::account::AccessKey;
 use crate::errors::TxExecutionError;
 use crate::hash::{hash, CryptoHash};
 use crate::logging;
@@ -40,11 +40,6 @@ impl Transaction {
     pub fn get_hash(&self) -> CryptoHash {
         let bytes = self.try_to_vec().expect("Failed to deserialize");
         hash(&bytes)
-    }
-
-    pub fn sign(self, signer: &dyn Signer) -> SignedTransaction {
-        let signature = signer.sign(self.get_hash().as_ref());
-        SignedTransaction::new(signature, self)
     }
 }
 
@@ -177,108 +172,6 @@ impl SignedTransaction {
 
     pub fn get_hash(&self) -> CryptoHash {
         self.hash
-    }
-
-    pub fn from_actions(
-        nonce: Nonce,
-        signer_id: AccountId,
-        receiver_id: AccountId,
-        signer: &dyn Signer,
-        actions: Vec<Action>,
-        block_hash: CryptoHash,
-    ) -> Self {
-        Transaction {
-            nonce,
-            signer_id,
-            public_key: signer.public_key(),
-            receiver_id,
-            block_hash,
-            actions,
-        }
-        .sign(signer)
-    }
-    pub fn send_money(
-        nonce: Nonce,
-        signer_id: AccountId,
-        receiver_id: AccountId,
-        signer: &dyn Signer,
-        deposit: Balance,
-        block_hash: CryptoHash,
-    ) -> Self {
-        Self::from_actions(
-            nonce,
-            signer_id,
-            receiver_id,
-            signer,
-            vec![Action::Transfer(TransferAction { deposit })],
-            block_hash,
-        )
-    }
-
-    pub fn stake(
-        nonce: Nonce,
-        signer_id: AccountId,
-        signer: &dyn Signer,
-        stake: Balance,
-        public_key: PublicKey,
-        block_hash: CryptoHash,
-    ) -> Self {
-        Self::from_actions(
-            nonce,
-            signer_id.clone(),
-            signer_id,
-            signer,
-            vec![Action::Stake(StakeAction { stake, public_key })],
-            block_hash,
-        )
-    }
-
-    pub fn create_account(
-        nonce: Nonce,
-        originator: AccountId,
-        new_account_id: AccountId,
-        amount: Balance,
-        public_key: PublicKey,
-        signer: &dyn Signer,
-        block_hash: CryptoHash,
-    ) -> Self {
-        Self::from_actions(
-            nonce,
-            originator,
-            new_account_id,
-            signer,
-            vec![
-                Action::CreateAccount(CreateAccountAction {}),
-                Action::AddKey(AddKeyAction {
-                    public_key,
-                    access_key: AccessKey { nonce: 0, permission: AccessKeyPermission::FullAccess },
-                }),
-                Action::Transfer(TransferAction { deposit: amount }),
-            ],
-            block_hash,
-        )
-    }
-
-    pub fn delete_account(
-        nonce: Nonce,
-        signer_id: AccountId,
-        receiver_id: AccountId,
-        beneficiary_id: AccountId,
-        signer: &dyn Signer,
-        block_hash: CryptoHash,
-    ) -> Self {
-        Self::from_actions(
-            nonce,
-            signer_id,
-            receiver_id,
-            signer,
-            vec![Action::DeleteAccount(DeleteAccountAction { beneficiary_id })],
-            block_hash,
-        )
-    }
-
-    pub fn empty(block_hash: CryptoHash) -> Self {
-        Self::from_actions(0, "".to_string(), "".to_string(), &EmptySigner {}, vec![], block_hash)
     }
 }
 
@@ -434,7 +327,7 @@ mod tests {
 
     use borsh::BorshDeserialize;
 
-    use near_crypto::{InMemorySigner, KeyType, Signature};
+    use near_crypto::{InMemorySigner, KeyType, Signature, Signer};
 
     use crate::account::{AccessKeyPermission, FunctionCallPermission};
     use crate::serialize::to_base;
